Back and working. Ported from audio-player project.
There are some problems with Avalonia's binding system. I'm seeing bindings that aren't being routed past the first control. So, the nested controls have to be set manually. I'm in need of an answer for this problem. Does it have to do with view location? I've already disabled the view locator (out of the box and otherwise).