Added some controls. Ready to switch from Avalonia, and finally switch to Java.
There are way too many bugs in Avalonia to promote it as a platform for a serious project. I think they took way too much on and can't keep up with development. The primary turn-off was binding bugs. The typical perk of WPF was that binding was almost always consistent. You may not be able to come back from binding issues when their backend is still being developed (!!!) I'd really rather look into their XAML compiler and see why these are happening; but they are very basic; and prevent use of the platform. Sorry guys, it was a good try.