Added radio-browser (radio-browser.info) online radio service
Mostly ready to stream radio EXCEPT for MSFT's HttpClient. It's hanging on every request call - and I can't help but think there's a reason that it does that from "the way I do it" Thread. So, if there's a synchronization context I haven't tried - it's pretty much the one where I rewrite the http client not to hang on a separate thread during retrieval. Synchronous "Send" also uses a Task in the background. So, this (BAD) design won't help unless there's a provided workaround.