Caved to Window's media player (for the backend) (because it worked OOB!)
There were PLENTY of problems working with Task-oriented HTTP streams. They were my problems to resolve; but there were no documents (OR) examples for what I was building. Why? Because people are used to things "just working". NAudio was going to work out; and I had it reading the "first frame" (for mp3 decoding) of "any radio-browser" stream. That being said, it would take the rest of the learning to know why there were still COM exceptions from NAudio. I would love to dig in; but by the time this is over, I will have already gotten down to the native portion. So, may as well "use what's there" right? Anyway, this player is just for my own use, probably. There will be another, Java, skywalker... (to be continued)