Reworked image caching per-entity type / also per TagLib PictureType. More memory control.
Most image caching will be for small images. The full sized images are ready to set for short-term memory only. Most of the time, the list virtualization handles things properly - so long as you are in control of the image sources (memory leaks!). Re-scaling the images and handling image sources was best accomplished using the GDI api (old winforms). The memory footprint target is <= 200 MB. So far, so good.