Import Basic Finished! (finally)
There was a refactoring of the library loader to put into place a simple workflow with steps. So, mostly have had to find time to work on these next issues with the library loader. The design is belabored by tedious UI issues, incomplete tag data, and dependency on 3rd party data that is also burdensome. So, after refactoring several times, we have a way to get to the final design for the import.