Application Startup:  Fixed a difficult TOI bug; setup bootstrapper for async / await dialog.
The target of invocation bug did not have a callstack. The actual bug was the logger - which did not have the proper parameter count from a recent change. There was no (local) trigger event that was throwing the exception, so the log must've been coming from somewhere else. Likely the database where there was some sort of exception thrown. Anyway, there have been a few sticking points using EF / Npgsql - but most of which are related to logging - because of all the backend complexity where we have a handle (logging event data).