Working re-work of library importer
Using Task / Dispatcher pattern. Added interfaces to the view model - which is a little messy at the moment. But, it will work out to manage the import this way. There's still a lot more to go on the importer to make it as automated as possible - while allowing lots of user interaction.